network: moving the IP Group parser up
